add Vitals section to Host details
