+++ frontend/pages/hosts/details/HostDetailsPage/_styles.scss
@@ -41,7 +41,7 @@
     }
   }
 
-  .about-card,
+  .vitals-card,
   .agent-options-card {
     .info {
       &__item {
