+++ frontend/pages/hosts/details/cards/HostSummary/HostSummary.tsx
@@ -8,36 +8,23 @@ import {
   isLinuxDiskEncryptionStatus,
 } from "interfaces/mdm";
 import { IOSSettings, IHostMaintenanceWindow } from "interfaces/host";
-import { IAppleDeviceUpdates } from "interfaces/config";
 import {
-  DiskEncryptionSupportedPlatform,
   isAndroid,
   isIPadOrIPhone,
   isDiskEncryptionSupportedLinuxPlatform,
   isOsSettingsDisplayPlatform,
-  platformSupportsDiskEncryption,
 } from "interfaces/platform";
-import { ROLLING_ARCH_LINUX_VERSIONS } from "interfaces/software";
 
 import getHostStatusTooltipText from "pages/hosts/helpers";
 
-import TooltipWrapperArchLinuxRolling from "components/TooltipWrapperArchLinuxRolling";
 import TooltipWrapper from "components/TooltipWrapper";
-import Icon from "components/Icon/Icon";
 import Card from "components/Card";
 import DataSet from "components/DataSet";
 import StatusIndicator from "components/StatusIndicator";
 import IssuesIndicator from "pages/hosts/components/IssuesIndicator";
-import DiskSpaceIndicator from "pages/hosts/components/DiskSpaceIndicator";
-import {
-  humanHostMemory,
-  wrapFleetHelper,
-  removeOSPrefix,
-  compareVersions,
-} from "utilities/helpers";
+
 import {
   DATE_FNS_FORMAT_STRINGS,
-  DEFAULT_EMPTY_CELL_VALUE,
 } from "utilities/constants";
 
 import OSSettingsIndicator from "./OSSettingsIndicator";
@@ -62,67 +49,9 @@ interface IHostSummaryProps {
   toggleOSSettingsModal?: () => void;
   toggleBootstrapPackageModal?: () => void;
   hostSettings?: IHostMdmProfile[];
-  osVersionRequirement?: IAppleDeviceUpdates;
   osSettings?: IOSSettings;
   className?: string;
 }
-
-const DISK_ENCRYPTION_MESSAGES = {
-  darwin: {
-    enabled: (
-      <>
-        The disk is encrypted. The user must enter their
-        <br /> password when they start their computer.
-      </>
-    ),
-    disabled: (
-      <>
-        The disk might be encrypted, but FileVault is off. The
-        <br /> disk can be accessed without entering a password.
-      </>
-    ),
-  },
-  windows: {
-    enabled: (
-      <>
-        The disk is encrypted. If recently turned on,
-        <br /> encryption could take awhile.
-      </>
-    ),
-    disabled: "The disk is unencrypted.",
-  },
-  linux: {
-    enabled: "The disk is encrypted.",
-    unknown: "The disk may be encrypted.",
-  },
-};
-
-const getHostDiskEncryptionTooltipMessage = (
-  platform: DiskEncryptionSupportedPlatform, // TODO: improve this type
-  diskEncryptionEnabled = false
-) => {
-  if (platform === "chrome") {
-    return "Fleet does not check for disk encryption on Chromebooks, as they are encrypted by default.";
-  }
-
-  if (
-    platform === "rhel" ||
-    platform === "ubuntu" ||
-    platform === "arch" ||
-    platform === "archarm" ||
-    platform === "manjaro" ||
-    platform === "manjaro-arm"
-  ) {
-    return DISK_ENCRYPTION_MESSAGES.linux[
-      diskEncryptionEnabled ? "enabled" : "unknown"
-    ];
-  }
-
-  // mac or windows
-  return DISK_ENCRYPTION_MESSAGES[platform][
-    diskEncryptionEnabled ? "enabled" : "disabled"
-  ];
-};
 
 const HostSummary = ({
   summaryData,
@@ -131,7 +60,6 @@ const HostSummary = ({
   toggleOSSettingsModal,
   toggleBootstrapPackageModal,
   hostSettings,
-  osVersionRequirement,
   osSettings,
   className,
 }: IHostSummaryProps): JSX.Element => {
@@ -141,11 +69,9 @@ const HostSummary = ({
     status,
     platform,
     os_version,
-    disk_encryption_enabled: diskEncryptionEnabled,
   } = summaryData;
 
   const isAndroidHost = isAndroid(platform);
-  const isChromeHost = platform === "chrome";
   const isIosOrIpadosHost = isIPadOrIPhone(platform);
 
   const renderIssues = () => (
@@ -176,179 +102,6 @@ const HostSummary = ({
       }
     />
   );
-
-  const renderDiskSpaceSummary = () => {
-    // Hide disk space field if storage measurement is not supported (sentinel value -1)
-    if (
-      typeof summaryData.gigs_disk_space_available === "number" &&
-      summaryData.gigs_disk_space_available < 0
-    ) {
-      return null;
-    }
-
-    const title = isAndroidHost ? (
-      <TooltipWrapper tipContent="Includes internal and removable storage (e.g. microSD card).">
-        Disk space
-      </TooltipWrapper>
-    ) : (
-      "Disk space"
-    );
-
-    return (
-      <DataSet
-        title={title}
-        value={
-          <DiskSpaceIndicator
-            gigsDiskSpaceAvailable={summaryData.gigs_disk_space_available}
-            percentDiskSpaceAvailable={summaryData.percent_disk_space_available}
-            gigsTotalDiskSpace={summaryData.gigs_total_disk_space}
-            gigsAllDiskSpace={summaryData.gigs_all_disk_space}
-            platform={platform}
-            tooltipPosition="bottom"
-          />
-        }
-      />
-    );
-  };
-  const renderDiskEncryptionSummary = () => {
-    if (!platformSupportsDiskEncryption(platform, os_version)) {
-      return <></>;
-    }
-    const tooltipMessage = getHostDiskEncryptionTooltipMessage(
-      platform,
-      diskEncryptionEnabled
-    );
-
-    let statusText;
-    switch (true) {
-      case isChromeHost:
-        statusText = "Always on";
-        break;
-      case diskEncryptionEnabled === true:
-        statusText = "On";
-        break;
-      case diskEncryptionEnabled === false:
-        statusText = "Off";
-        break;
-      case (diskEncryptionEnabled === null ||
-        diskEncryptionEnabled === undefined) &&
-        platformSupportsDiskEncryption(platform, os_version):
-        statusText = "Unknown";
-        break;
-      default:
-        // something unexpected happened on the way to this component, display whatever we got or
-        // "Unknown" to draw attention to the issue.
-        statusText = diskEncryptionEnabled || "Unknown";
-    }
-
-    return (
-      <DataSet
-        title="Disk encryption"
-        value={
-          <TooltipWrapper tipContent={tooltipMessage}>
-            {statusText}
-          </TooltipWrapper>
-        }
-      />
-    );
-  };
-
-  const renderOperatingSystemSummary = () => {
-    // No tooltip if minimum version is not set, including all Windows, Linux, ChromeOS, Android operating systems
-    if (!osVersionRequirement?.minimum_version) {
-      const version = summaryData.os_version;
-      const versionForRender = ROLLING_ARCH_LINUX_VERSIONS.includes(version) ? (
-        // wrap a tooltip around the "rolling" suffix
-        <>
-          {version.slice(0, -8)}
-          <TooltipWrapperArchLinuxRolling />
-        </>
-      ) : (
-        version
-      );
-      return (
-        <DataSet
-          title="Operating system"
-          value={versionForRender}
-          className={`${baseClass}__os-data-set`}
-        />
-      );
-    }
-
-    const osVersionWithoutPrefix = removeOSPrefix(summaryData.os_version);
-    const osVersionRequirementMet =
-      compareVersions(
-        osVersionWithoutPrefix,
-        osVersionRequirement.minimum_version
-      ) >= 0;
-
-    return (
-      <DataSet
-        title="Operating system"
-        value={
-          <>
-            {!osVersionRequirementMet && (
-              <Icon name="error-outline" color="ui-fleet-black-75" />
-            )}
-            <TooltipWrapper
-              tipContent={
-                osVersionRequirementMet ? (
-                  "Meets minimum version requirement."
-                ) : (
-                  <>
-                    Does not meet minimum version requirement.
-                    <br />
-                    Deadline to update: {osVersionRequirement.deadline}
-                  </>
-                )
-              }
-            >
-              {summaryData.os_version}
-            </TooltipWrapper>
-          </>
-        }
-      />
-    );
-  };
-
-  const renderAgentSummary = () => {
-    if (isIosOrIpadosHost || isAndroidHost) {
-      return null;
-    }
-
-    if (isChromeHost) {
-      return <DataSet title="Agent" value={summaryData.osquery_version} />;
-    }
-
-    if (summaryData.orbit_version !== DEFAULT_EMPTY_CELL_VALUE) {
-      return (
-        <DataSet
-          title="Agent"
-          value={
-            <TooltipWrapper
-              tipContent={
-                <>
-                  osquery: {summaryData.osquery_version}
-                  <br />
-                  Orbit: {summaryData.orbit_version}
-                  {summaryData.fleet_desktop_version !==
-                    DEFAULT_EMPTY_CELL_VALUE && (
-                    <>
-                      <br />
-                      Fleet Desktop: {summaryData.fleet_desktop_version}
-                    </>
-                  )}
-                </>
-              }
-            >
-              {summaryData.orbit_version}
-            </TooltipWrapper>
-          }
-        />
-      );
-    }
-    return <DataSet title="Osquery" value={summaryData.osquery_version} />;
-  };
 
   const renderMaintenanceWindow = ({
     starts_at,
@@ -439,12 +192,7 @@ const HostSummary = ({
           }
         />
       )}
-      {summaryData.issues?.total_issues_count > 0 &&
-        !isIosOrIpadosHost &&
-        !isAndroidHost &&
-        renderIssues()}
       {isPremiumTier && renderHostTeam()}
-      {/* Rendering of OS Settings data */}
       {isOsSettingsDisplayPlatform(platform, os_version) &&
         hostSettings &&
         hostSettings.length > 0 && (
@@ -458,6 +206,10 @@ const HostSummary = ({
             }
           />
         )}
+      {summaryData.issues?.total_issues_count > 0 &&
+        !isIosOrIpadosHost &&
+        !isAndroidHost &&
+        renderIssues()}
       {bootstrapPackageData?.status && !isIosOrIpadosHost && !isAndroidHost && (
         <DataSet
           title="Bootstrap package"
@@ -469,19 +221,6 @@ const HostSummary = ({
           }
         />
       )}
-      {!isChromeHost && renderDiskSpaceSummary()}
-      {renderDiskEncryptionSummary()}
-      {!isIosOrIpadosHost && (
-        <DataSet
-          title="Memory"
-          value={wrapFleetHelper(humanHostMemory, summaryData.memory)}
-        />
-      )}
-      {!isIosOrIpadosHost && (
-        <DataSet title="Processor type" value={summaryData.cpu_type} />
-      )}
-      {renderOperatingSystemSummary()}
-      {renderAgentSummary()}
       {isPremiumTier &&
         // TODO - refactor normalizeEmptyValues pattern
         !!summaryData.maintenance_window &&
