+++ frontend/pages/hosts/details/cards/Vitals/_styles.scss
@@ -1,8 +1,8 @@
-.about-card {
+.vitals-card {
   @include vertical-card-layout;
 
   .truncated-tooltip {
-    .about-card__device-mapping__source {
+    .vitals-card__device-mapping__source {
       color: inherit;
     }
   }
@@ -44,12 +44,12 @@
   // TooltipTruncatedText component.
   &__info-grid {
     display: grid;
-    grid-template-columns: repeat(2, minmax(150px, max-content));
+    grid-template-columns: repeat(auto-fit, minmax(150px, max-content));
     gap: $gap-data-sets $pad-xxlarge;
 
-    // at the larger screen widths we want to have 3 columns.
+    // at the larger screen widths we want to have 5 columns.
     @media (min-width: $break-xxl) {
-      grid-template-columns: repeat(3, minmax(150px, max-content));
+      grid-template-columns: repeat(5, minmax(150px, max-content));
     }
   }
