+++ frontend/utilities/constants.tsx
@@ -408,25 +408,14 @@ export const HOST_SUMMARY_DATA: (keyof IHost)[] = [
   "id",
   "status",
   "issues",
-  "memory",
-  "cpu_type",
   "platform",
-  "os_version",
-  "osquery_version",
-  "orbit_version",
-  "fleet_desktop_version",
   "detail_updated_at",
-  "percent_disk_space_available",
-  "gigs_disk_space_available",
-  "gigs_total_disk_space",
-  "gigs_all_disk_space",
   "team_name",
-  "disk_encryption_enabled",
   "display_name", // Not rendered on my device page
   "maintenance_window", // Not rendered on my device page
 ];
 
-export const HOST_ABOUT_DATA = [
+export const HOST_VITALS_DATA = [
   "seen_time",
   "uptime",
   "last_enrolled_at",
@@ -440,6 +429,17 @@ export const HOST_ABOUT_DATA = [
   "last_restarted_at",
   "platform",
   "uuid",
+  "gigs_disk_space_available",
+  "percent_disk_space_available",
+  "gigs_total_disk_space",
+  "gigs_all_disk_space",
+  "disk_encryption_enabled",
+  "osquery_version",
+  "orbit_version",
+  "fleet_desktop_version",
+  "memory",
+  "cpu_type",
+  "os_version",
 ];
 
 export const HOST_OSQUERY_DATA = [
